% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrology_Helper.R
\name{get_wateryear}
\alias{get_wateryear}
\title{Determine Water Year}
\usage{
get_wateryear(date, yearstart)
}
\arguments{
\item{date}{The date in the form of date or datetime.}

\item{yearstart}{The month your water year starts on, defaults as Oct 1st (10).}
}
\value{
The water year the date belongs to.
}
\description{
This function determines the water year based on a date given
}
\examples{
get_wateryear(lubridate::today(),9)
df<- dplyr::mutate(df, wateryear = get_wateryear(datetime))
}
