% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrology_Helper.R
\name{parse_dates}
\alias{parse_dates}
\title{Identify what storms an observation belongs to}
\usage{
parse_dates(df, df2)
}
\arguments{
\item{df}{The dataframe to manipulate. You need datetime at minimum.}

\item{df2}{The dataframe output from parse_storms() function.}
}
\value{
Returns updated dataframe where each observation has an associated
}
\description{
This function determines what storm event an observation is associated with.
}
\examples{
df<- parse(df, df2)
}
