% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrology_Helper.R
\name{parse_storms}
\alias{parse_storms}
\title{Parse storm events from precipitation data}
\usage{
parse_storms(df, intervals_per_hr, interevent_period_hr, storm_size_minimum)
}
\arguments{
\item{df}{This is a dataframe that contains datetime columns and a rainfall
columns labeled as "depth_in" or "depth_mm".}

\item{intervals_per_hr}{How many recordings per hour. 1 means hourly,}

\item{interevent_period_hr}{The dry period needed in between storm events, in hours}

\item{storm_size_minimum}{What is the minimum storm size being evaluated?}
}
\value{
A new data frame with storm_id, water year,eventstart, eventend,
total_depth, and duration_hr
}
\description{
This function parses storm events from tidy precipitation data.
}
\examples{
df<- parse_storms (df = df,intervals_per_hr = 1, interevent_period_hr = 12,storm_size_minimum = 0.02)
}
