% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outputting.R
\name{report_compiler}
\alias{report_compiler}
\title{Render markdown report}
\usage{
report_compiler(input_location, input_name, output_location, output_name)
}
\arguments{
\item{input_location}{Location of R script, in quotes.}

\item{input_name}{Script name , in quotes.}

\item{output_location}{Location where markdown will be saved, in quotes.}

\item{output_name}{File name, in quotes}
}
\value{
Saved markdown file.
}
\description{
This function renders and saves a markdown directly from a R script
to any desired location. `
}
\examples{
report_compiler(input_location = "scripts",input_namne = "mymarkdownreport.R",output_location = "output",output_name = "Report.html")
}
